+++ shardingsphere-infra/shardingsphere-infra-executor/src/main/java/org/apache/shardingsphere/infra/executor/sql/raw/execute/result/update/ExecuteUpdateResult.java
@@ -15,37 +15,20 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.executor.sql.raw.execute.result.impl;
+package org.apache.shardingsphere.infra.executor.sql.raw.execute.result.update;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
+import org.apache.shardingsphere.infra.executor.sql.raw.execute.result.ExecuteResult;
 
 /**
- * Query header.
+ * Execute update result.
  */
 @RequiredArgsConstructor
 @Getter
-public final class QueryHeader {
+public final class ExecuteUpdateResult implements ExecuteResult {
     
-    private final String schema;
+    private final int updateCount;
     
-    private final String table;
-    
-    private final String columnLabel;
-    
-    private final String columnName;
-    
-    private final int columnLength;
-    
-    private final Integer columnType;
-    
-    private final int decimals;
-    
-    private final boolean signed;
-    
-    private final boolean primaryKey;
-    
-    private final boolean notNull;
-    
-    private final boolean autoIncrement;
+    private final long lastInsertId;
 }
